chore: set up right rollup configurations
re #144
