+++ packages/baby-jubjub/src/types/index.ts
@@ -1,3 +1,3 @@
-import { BigNumber } from "@zk-kit/utils"
+import type { BigNumber } from "@zk-kit/utils"
 
 export type Point<N = BigNumber> = [N, N]
