+++ packages/eddsa-poseidon/rollup.config.ts
@@ -1,8 +1,10 @@
 import commonjs from "@rollup/plugin-commonjs"
 import { nodeResolve } from "@rollup/plugin-node-resolve"
+import terser from "@rollup/plugin-terser"
+import typescript from "@rollup/plugin-typescript"
 import fs from "fs"
 import cleanup from "rollup-plugin-cleanup"
-import typescript from "@rollup/plugin-typescript"
+import nodePolyfills from "rollup-plugin-polyfill-node"
 
 const pkg = JSON.parse(fs.readFileSync("./package.json", "utf8"))
 const banner = `/**
@@ -14,19 +16,53 @@ const banner = `/**
  * @see [Github]{@link ${pkg.homepage}}
 */`
 
-export default {
-    input: "src/index.ts",
-    output: [
-        { file: pkg.exports["."].require, format: "cjs", banner },
-        { file: pkg.exports["."].default, format: "es", banner }
-    ],
-    external: Object.keys(pkg.dependencies),
-    plugins: [
-        typescript({ tsconfig: "./build.tsconfig.json" }),
-        commonjs(),
-        nodeResolve({
-            preferBuiltins: true
-        }),
-        cleanup({ comments: "jsdoc" })
-    ]
-}
+const name = pkg.name.split("/")[1].replace(/[-/]./g, (x: string) => x.toUpperCase()[1])
+
+export default [
+    {
+        input: "src/index.ts",
+        output: [
+            { file: pkg.exports["."].require, format: "cjs", banner },
+            { file: pkg.exports["."].default, format: "es", banner }
+        ],
+        external: [
+            ...Object.keys(pkg.dependencies),
+            "@zk-kit/utils/conversions",
+            "@zk-kit/utils/f1-field",
+            "@zk-kit/utils/scalar",
+            "@zk-kit/utils/error-handlers",
+            "@zk-kit/utils/type-checks"
+        ],
+        plugins: [
+            typescript({ tsconfig: "./build.tsconfig.json" }),
+            nodeResolve(),
+            commonjs(),
+            cleanup({ comments: "jsdoc" })
+        ]
+    },
+    {
+        input: "src/index.ts",
+        output: [
+            {
+                file: pkg.iife,
+                name,
+                format: "iife",
+                banner
+            },
+            {
+                file: pkg.unpkg,
+                name,
+                format: "iife",
+                plugins: [terser({ output: { preamble: banner } })]
+            }
+        ],
+        external: [],
+        plugins: [
+            typescript({ tsconfig: "./build.tsconfig.json" }),
+            nodeResolve({ preferBuiltins: false, browser: true }),
+            commonjs(),
+            nodePolyfills(),
+            cleanup({ comments: "jsdoc" })
+        ]
+    }
+]
