+++ packages/eddsa-poseidon/src/types/index.ts
@@ -1,5 +1,5 @@
-import { Point } from "@zk-kit/baby-jubjub"
-import { BigNumber } from "@zk-kit/utils"
+import type { Point } from "@zk-kit/baby-jubjub"
+import type { BigNumber } from "@zk-kit/utils"
 
 export type Signature<N = BigNumber> = {
     R8: Point<N>
