+++ packages/utils/tsconfig.json
@@ -1,4 +1,4 @@
 {
     "extends": "../../tsconfig.json",
-    "include": ["src", "tests", "rollup.node.config.ts", "rollup.browser.config.ts"]
+    "include": ["src", "tests", "rollup.config.ts"]
 }
